requirements.txt
aiohttp==3.9.1
aiosignal==1.3.1
anyio==4.2.0
apispec==6.3.1
async-timeout==4.0.3
attrs==23.1.0
bech32==1.2.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.7
cosmpy==0.9.1
discord==2.3.2
discord.py==2.3.2
distro==1.9.0
dnspython==2.4.2
ecdsa==0.18.0
exceptiongroup==1.2.0
frozenlist==1.4.1
googleapis-common-protos==1.62.0
grpcio==1.51.1
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
msgpack==1.0.7
multidict==6.0.4
openai==1.6.1
packaging==23.2
protobuf==4.25.1
pycryptodome==3.19.1
pydantic==1.10.13
pymongo==4.6.1
python-dateutil==2.8.2
python-dotenv==1.0.0
redis==5.0.1
referencing==0.32.0
requests==2.31.0
rpds-py==0.16.2
six==1.16.0
sniffio==1.3.0
tqdm==4.66.1
typing_extensions==4.9.0
uagents==0.9.2
urllib3==2.1.0
uvicorn==0.20.0
websockets==10.4
yarl==1.9.4
